pid double to int

+++ lib/utsav/payment_screen.dart
@@ -1487,7 +1487,7 @@ class _PaymentScreenState extends State<PaymentScreen> {
                                     }
                                 }
                               Map<String,dynamic> product={
-                                "product_id":double.parse(cart.cartItem[index].productId.toString()),
+                                "product_id":int.parse(cart.cartItem[index].productId.toString()),
                                 "variation_id":double.parse(variation![index]),
                                 "quantity": cart.cartItem[index].quantity,
                                 "unit_price": cart.cartItem[index].unitPrice,
